'余额显示更改'

+++ app/Http/Controllers/Admin/SchoolController.php
@@ -629,7 +629,7 @@ class SchoolController extends Controller
         $school = json_decode(json_encode($school),true);
         $school['ifinto'] = $school['ifinto']>0?true:false;//
         $school['recharge_balance'] = $school['balance'];//充值余额
-        $school['total_balance'] = (int) $school['balance'] + (int) $school['give_balance'];//全部余额
+        $school['total_balance'] = (float) $school['balance'] + (float) $school['give_balance'];//全部余额
 
         //管理员信息
         $field = ['username','realname','mobile'];
